requirements.txt
streamlit==1.32.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
numpy==1.26.4
pandas==2.1.4
nltk==3.8.1
regex==2023.12.25
fpdf2==2.7.5
python-docx==0.8.11
PyPDF2==3.0.1
plotly==5.18.0
tqdm==4.66.1
scipy==1.11.4
torch==2.1.0
transformers==4.35.2
protobuf==4.25.1
joblib==1.3.2